Closes 02: "Implementa pagina de chat e melhora home."
