+++ pages/chat.js
@@ -1,0 +1,5 @@
+export default function ChatPage() {
+  return (
+    <div>Página do Chat</div>
+  )
+}
